feat: create item Card
