+++ src/pages/SearchPage.vue
@@ -38,45 +38,18 @@
       <ArticleSkeleton />
     </section>
     <section v-if="promptStore.getPrompts.filter((prompt) => prompt.categories.includes(category)).length">
-      <article
-        v-for="prompt in prompts"
-        class="q-pt-md relative-position row"
-        :key="prompt?.id"
-        v-ripple:primary
-        v-show="prompt.categories.includes(category)"
-        @click="goToPrompt(prompt.slug)"
-      >
-        <div class="col-8">
-          <div class="flex items-center">
-            <q-avatar size="2rem">
-              <q-img :src="prompt.author.photoURL" />
-            </q-avatar>
-            <p class="q-mb-none q-ml-sm text-body1">
-              {{ prompt.author.displayName.length > 20 ? prompt.author.displayName.substring(0, 20) + '...' : prompt.author.displayName }}
-            </p>
-          </div>
-          <h2 class="q-mb-none text-body1 text-bold">
-            {{ prompt.title.length > 40 ? prompt.title.substring(0, 40) + '...' : prompt.title }}
-          </h2>
-          <p class="q-my-none text-body2 text-secondary">
-            {{ shortMonthDay(prompt.created) }} &nbsp;•&nbsp; {{ giveReadingTime(prompt.description) }}min
-          </p>
-          <div v-if="category === 'Trending'">
-            <q-badge v-for="(item, i) of prompt.categories" class="q-mx-xs" :key="i" rounded>{{ item }}</q-badge>
-          </div>
-        </div>
-        <q-img class="col-4" :ratio="1" :src="prompt.image" spinner-color="primary" spinner-size="3rem" style="border-radius: 24px" />
-        <!-- TODO: Add 'Selected for you' and two more buttons according to mockup -->
-        <q-separator class="full-width q-mt-md" />
-      </article>
+
+      <ItemCard v-for="prompt in prompts" :key="prompt?.id" :item="prompt"
+                v-show="prompt.categories.includes(category) || category === 'All'"
+                :link="`/prompt/${prompt.slug}`"></ItemCard>
     </section>
   </q-page>
 </template>
 
 <script setup>
 import ArticleSkeleton from 'src/components/ArticleSkeleton.vue'
+import ItemCard from "components/ItemCard.vue";
 import { usePromptStore } from 'src/stores'
-import { shortMonthDay } from 'src/utils/date'
 import { onMounted, ref } from 'vue'
 import { useRouter } from 'vue-router'
 
@@ -105,23 +78,4 @@ onMounted(async () => {
   prompts.value = promptStore.getPrompts
 })
 
-function goToPrompt(slug) {
-  router.push(`/prompt/${slug}`)
-}
-
-function giveReadingTime(description) {
-  let stringArr = description.split(' ')
-  let adjustment = 0
-  stringArr = stringArr.filter((el) => {
-    if (el.indexOf('<') == -1) {
-      return el
-    }
-    adjustment++
-  })
-
-  let wordCount = stringArr.length + adjustment
-  const averageWordsPerMinute = 238
-
-  return Math.round(wordCount / averageWordsPerMinute)
-}
 </script>
